changed test location
